Check for args with isNaN() instead of undefined
With the change to the aliased arguments, both startRevision and
stopRevision are now defined through parseInt(), but are 'NaN' if
not specifically passed.

Use isNaN() to check for valid values, and print usage instructions
if one or both are not set. Note: using 'typeof' number will return
true, because parseInt() returns a number object, even though it's
not a real number. Because JS.